requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1

